LayoutTests/editing/spelling/context-menu-suggestions.html

<body onload="test()">
<p>This test checks that spelling suggestions in the context menu are not show when spellcheck="false".
To test manually, right click on the text in the input box.</p>
<input id="input" spellcheck="false" />
</body>
<script>
function test()
{
    if (window.layoutTestController)
        layoutTestController.dumpAsText();

    var input = document.getElementById("input");
    input.focus();
    document.execCommand("InsertText", false, 'w');
    document.execCommand("InsertText", false, 'w');
    document.execCommand("InsertText", false, ' ');

    var x = input.offsetParent.offsetLeft + input.offsetLeft + 4;
    var y = input.offsetParent.offsetTop + input.offsetTop + input.offsetHeight / 2;

    if (!window.eventSender)
        return;

    console.log("Spelling suggestings should NOT be here");
    eventSender.mouseMoveTo(x, y);
    eventSender.contextClick(true);

    input.setAttribute("spellcheck", "true");

    console.log("Spelling suggestings SHOULD be here");
    eventSender.mouseMoveTo(x, y);
    eventSender.contextClick(true);
}
</script>
